Removed skyboxes from renderering for now as we transition over to merging HDRimages and images together.
