+++ src/graphics/material/image.rs
@@ -1,9 +1,12 @@
+use std::{fs, io};
+
 pub struct Image {
     pub(crate) name: String,
     pub(crate) texture: wgpu::Texture,
     pub(crate) extent: wgpu::Extent3d,
     pub(crate) sampler: wgpu::Sampler,
     pub(crate) view: wgpu::TextureView,
+    pub(crate) format: wgpu::TextureFormat,
 }
 
 impl Image {
@@ -17,14 +20,11 @@ impl Image {
         T: Into<String>,
     {
         let path = path.into();
-        let img = image::open(&path)
-            .unwrap_or_else(|_| panic!("Image: Unable to open the file: {}", path))
-            .to_rgba();
-        let (width, height) = img.dimensions();
-        let texture_extent = wgpu::Extent3d {
-            width,
-            height,
-            depth: 1,
+
+        let (image_bytes, texture_extent, format) = if path.ends_with(".hdr") {
+            Self::create_hdr_image(path)
+        } else {
+            Self::create_normal_image(path)
         };
 
         let texture = device.create_texture(&wgpu::TextureDescriptor {
@@ -32,19 +32,18 @@ impl Image {
             mip_level_count: 1,
             sample_count: 1,
             dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::Rgba8UnormSrgb,
+            format,
             usage: wgpu::TextureUsage::SAMPLED | wgpu::TextureUsage::COPY_DST,
             label: None,
         });
 
-        let image_bytes: Vec<u8> = img.into_raw();
         let temp_buf = device.create_buffer_with_data(&image_bytes, wgpu::BufferUsage::COPY_SRC);
 
         encoder.copy_buffer_to_texture(
             wgpu::BufferCopyView {
                 buffer: &temp_buf,
                 offset: 0,
-                bytes_per_row: 4 * width,
+                bytes_per_row: if format == wgpu::TextureFormat::Rgba8UnormSrgb { 4 * texture_extent.width } else { (4 * 4) * texture_extent.width },
                 rows_per_image: 0,
             },
             wgpu::TextureCopyView {
@@ -76,6 +75,52 @@ impl Image {
             extent: texture_extent,
             sampler,
             view,
+            format,
         }
+    }
+
+    fn create_normal_image(path: String) -> (Vec<u8>, wgpu::Extent3d, wgpu::TextureFormat) {
+        let img = image::open(&path)
+            .unwrap_or_else(|_| panic!("Image: Unable to open the file: {}", path))
+            .to_rgba();
+        let (width, height) = img.dimensions();
+        let texture_extent = wgpu::Extent3d {
+            width,
+            height,
+            depth: 1,
+        };
+
+        let image_bytes: Vec<u8> = img.into_raw();
+
+        (image_bytes, texture_extent, wgpu::TextureFormat::Rgba8UnormSrgb)
+    }
+
+    fn create_hdr_image(path: String) -> (Vec<u8>, wgpu::Extent3d, wgpu::TextureFormat) {
+        // Load the image
+        let decoder =
+            image::hdr::HdrDecoder::new(io::BufReader::new(fs::File::open(&path).unwrap()))
+                .unwrap();
+        let metadata = decoder.metadata();
+        let decoded = decoder.read_image_hdr().unwrap();
+
+        let (w, h) = (metadata.width, metadata.height);
+
+        let texture_extent = wgpu::Extent3d {
+            width: w,
+            height: h,
+            depth: 1,
+        };
+
+        let image_data = decoded
+            .iter()
+            .flat_map(|pixel| vec![pixel[0], pixel[1], pixel[2], 1.0])
+            .collect::<Vec<_>>();
+        
+        let image_bytes = unsafe {
+            std::slice::from_raw_parts(image_data.as_ptr() as *const u8, image_data.len() * 4)
+        }
+        .to_vec();
+        
+        (image_bytes, texture_extent, wgpu::TextureFormat::Rgba32Float)
     }
 }
